Created new services. Removed old classes of services, entities, repositories.

+++ src/main/java/by/intexsoft/forum/repository/UserRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
-public interface UserRepository  extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 	User findByEmail(String email);
 }
-
