+++ src/main/java/by/intexsoft/forum/service/AbstractEntityService.java
@@ -5,11 +5,11 @@ import by.intexsoft.forum.entity.AbstractEntity;
 import java.util.List;
 
 public interface AbstractEntityService<T extends AbstractEntity> {
-	void delete(int id);
+	void delete(long id);
 
 	T save(T entity);
 
-	T find(int id);
+	T find(long id);
 
 	List<T> findAll();
 }
